backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0

# Async DB
asyncpg==0.30.0
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3

# Redis / caching / task queue
redis[hiredis]==5.1.0
celery[redis]==5.4.0
kombu==5.3.4

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# HTTP client
httpx==0.27.2
google-generativeai==0.8.3     # Gemini API — replaces Ollama
geopy==2.4.1          # Geodesic calculations (Haversine + PostGIS future)

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# ML / AI core
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.1
joblib==1.4.2

# Audio processing
librosa==0.10.2
soundfile==0.12.1
torchaudio==2.6.0

# Vision
torch==2.6.0
torchvision==0.21.0
Pillow==10.4.0

# NLP
spacy==3.8.14
langdetect==1.0.9
# medspacy removed — no release is compatible with spaCy 3.7.x (medspacy tops out at 3.5).
# The extractor uses en_core_web_sm for NER, which covers all 133 canonical symptoms.
# spaCy model installed via: python -m spacy download en_core_web_sm
# Add to Dockerfile: RUN python -m spacy download en_core_web_sm

# Speech-to-text (Whisper via HuggingFace Transformers)
transformers==4.44.2
accelerate==0.34.2

# Audio format conversion + noise reduction
pydub==0.25.1          # format conversion: ogg/webm/mp3 → wav
noisereduce==3.0.2     # spectral gating noise reduction for phone recordings
# soundfile already included above — WAV I/O

# ffmpeg is required by pydub for non-WAV formats
# In Docker: apt-get install -y ffmpeg (already in Dockerfile)

# Google Auth (FCM HTTP v1 API — replaces deprecated Legacy Server Key)
google-auth==2.32.0
google-auth-httplib2==0.2.0

# Observability
sentry-sdk[fastapi]==2.14.0
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
orjson==3.10.7
pytz==2024.2

# Analytics module dependencies (numpy/scipy already pinned above)
scipy==1.14.1

# PyTorch CPU-only wheels — avoids 2.5GB+ CUDA download and keeps build < 5 min
# Install with: pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cpu
